avoid unnecessary factorying of FieldProxyTarget
